requirements.txt
opencv-python==4.8.0.76
numpy==1.24.3
scikit-learn==1.3.0
pandas==2.0.3
pillow==10.0.0
matplotlib==3.7.2
flask==2.3.3 